Fix pagination bug when listing practitioner
The practitioners are a joined set of claim admins and enrollment officers.
When listing them, the result is paginated. To do so, they are merged and
iterated over.

However, with the demo data, we obtained duplicates and an IndexError in the
last page.

The duplicates are due to a wrong slice of the sets: it slices until the last
index that has been listed in the previous page rather than from.

The IndexError is due to the fact that the last index of the last step will be
always out of the bounds used to return the right set, i.e. here the last one.

This does not yet include tests to reproduce the bug. The bug has been produced
by iterating over all the pages when listing practitioner with the demo data.